dtpc_send: Use correct type for SDR object length
dtpc(3) says that the `length` parameter of dtpc_send() is the length of
an SDR object. The SDR dataspace is limited by size_t, and object
lengths returned by sdr_object_length() are also size_t. It's strange
that dtpc_send() instead types `length` as `unsigned`. Even stranger,
dtpc_send() relies on insertRecord() to insert the data item into an
outbound ADU, and insertRecord() instead types `length` as `signed`.

Fix dtpc_send() and insertRecord() to consistently use size_t.

estimateLength(), which is only used by insertRecord() for estimating
the length of the outbound ADU, trusts the SDNV lengths in the payload
records. Since insertRecord() is used on the outbound path, it's
unlikely that these lengths are wrong or will overflow, but we might as
well be defensive. If we do trust the SDNVs though, size_t for
totalLength is fine because of the aforementioned SDR dataspace limit.
But the SDR cap should be lifted eventually, so uvast is appropriate.

We want to follow IWYU, so also include the right headers in the
affected files. Our codebase would like to include system headers before
project headers, but since ION's headers define feature test macros, we
need to include stddef.h after ION's headers.

+++ dtpc/include/dtpc.h
@@ -14,6 +14,10 @@
 #define DTPC_H
 
 #include "bp.h"
+#include "platform.h"
+#include "sdrxn.h"
+
+#include <stddef.h>
 
 #ifdef __cplusplus
 extern "C" {
@@ -76,7 +80,7 @@ extern int      dtpc_send(unsigned int profileID,
 			char *reportToEid,
 			int classOfService,
 			Object item,
-			unsigned int length);
+			size_t length);
 
 extern int      dtpc_receive(DtpcSAP sap,
 			DtpcDelivery *dlv,
